Android TextInput: Hide default native underline
By default, Android's native EditText control shows an underline. In RN, setting the TextInput's background color used to cause the underline to be hidden. In a newer version of RN, it doesn't and we must explicitly set the underline's color to transparen to hide it.

You can see screenshots with and without the underline at https://github.com/facebook/react-native/issues/5424#issuecomment-262425480.