Fix Button's highlight when it is pressed
When a Button is pressed, its opacity should change. However, this broke in React Native 0.42. The problem is that Button incorporates ReactNative's Touchable.Mixin. In between RN 0.34 and 0.42, Touchable.Mixin started implementing the methods `componentDidMount` and `componentWillUnmount`. These ended up overriding Button's implementations so Button never set _isMounted to true, and functionality like changing the Button's opacity when it is pressed broke.

I don't know of a good fix for this. React has a mixin mechanism which would avoid this problem but it isn't supported when using ES6 classes. Consequently, we have to rely on our own mixin mechanism. I handled the naming collisions manually and wrote some asserts so we can detect new collisions in the future.

I deleted MixinUtil.ts to discourage people from using mixins in other places.