UWP: Send correct range in TextInput's selection event (#425)
TextInput's onSelectionChange handler in RX UWP was attempting to clamp the selection range to be within the range of the TextInput's value. However, this wasn't working properly because it was clamping using an out-of-date length for the TextInput's value. The length was out-of-date because in React Native UWP, onSelectionChange fires before onChangeText (similar to React Native iOS). The impact was that the TextInput's cursor could be drawn in the wrong spot because the app would be setting the cursor position using the incorrect selection range that RX gave it.

    The fix is to avoid clamping on UWP similar to what is done on iOS.

    The clamping is to avoid a native exception due to using an out of bounds selection range but we've only seen such exceptions on Android.