Sync layout: Defer initial CustomScrollbar sizing
Asking for viewport dimensions triggers layout recalculation and
unnecessarily stalls JS execution. Scrollbars having the exact correct
size on first render is arguably not worth the perf hit (much less
so when they are initially hidden which is the common case).

Also deferring _tryLtrOverride() for the same reason. It's not clear
when the neutral override class would actually be used. Maybe this
whole thing can be removed? (Apologies, my web-fu is weak.)