Implement hideable popups
Popups that are lightweight, expected to be shown repeatedly, and easy to reset to
their initial state are good candidates for this feature. To elect to be "hideable",
a popup sets hideOnDismiss to true and optionally subscribes to onReshow().

When the popup is dismissed, instead of removing it from the DOM, it is still being
rendered, only with visibility: hidden. Then when the same popup is brought up again,
its onReshow() callback is called to give it a chance to reset its state, and visibility
goes back to its default value.

This massively improves the time required to bring up the popup as it saves a bunch of
work done by React when a new node subtree is introduced.