Implement cacheable popups (#534)
* Introduce PopupContainer

A new component to host the active popup and keep track of subcomponents rendered
into it.

* Wire up View to register with PopupContainer

Views that are part of the active popup need to be notified of its hiding and re-showing
for accessibility/focusing reasons. This is anologous to componentDidMount() and
componentWillUnmount() currently containing this logic for the regular case where the
popup is completely removed from the DOM.

* Implement hideable popups

Popups that are lightweight, expected to be shown repeatedly, and easy to reset to
their initial state are good candidates for this feature. To elect to be "hideable",
a popup sets hideOnDismiss to true and optionally subscribes to onReshow().

When the popup is dismissed, instead of removing it from the DOM, it is still being
rendered, only with visibility: hidden. Then when the same popup is brought up again,
its onReshow() callback is called to give it a chance to reset its state, and visibility
goes back to its default value.

This massively improves the time required to bring up the popup as it saves a bunch of
work done by React when a new node subtree is introduced.

* Feedback: Extend CommonProps and clone props before modifying

* Feedback: Remove onReshow and rename hideOnDismiss to cacheable

* Add popup caching on web (unlimited)

* Add popup caching on native (unlimited)

* Implement popup cache limits

Somewhat arbitrary hard-coded numbers.

* Update RXPTest and documentation

* Feedback: Subscribe to memoryWarningEvent and render fewer elements

* Feedback: Fix show/re-show logic on web and use named type for the tuple

* Feedback: Re-render RootView after releasing cached popups and indentation