Add Image.prefetchAndGetCachedPath
When prefetching an image, enables consumers to find out where on disk the image gets cached to. Useful if you need a local file path for an image rather than a URL and if you're okay with there being no guarantees around how long the local path will exist.

`RX.Image.prefetchAndGetCachedPath` is currently only interesting on Windows. On other platforms, it has the same effect as `RX.Image.prefetch`.

We've only invested in implementing this on Windows so far because, in our app, this API is needed for a Windows-specific scenario which requires calling a native API which accepts an image as a local file path but not as a URL. We can extend this API to other platforms in the future if needed.

React Native Windows added support for `prefetchAndGetCachedPath` in this commit: https://github.com/Microsoft/react-native-windows/commit/e0c0d820d5a932d9063be47537c47688ee6357e9