Fix focus management for cacheable popups on RN Windows
Just like web, Windows views rendered into a cacheable popup must be aware of the popup
hiding and re-showing in order to properly un-restrict and re-restrict focus to the
right components. The implementation is identical to that used on web, i.e. FocusManager
is released and re-initialized on hide and re-show, respectively, in addition to doing
it on unmount and mount.