Emulate Cache-Control: max-stale on iOS
iOS does not seem to respect the max-stale header. It does not load
expired images from the cache unless the force-cache option (internally
NSURLRequestReturnCacheDataElseLoad) is passed.

We can't use force-cache always because it has the opposite problem of
loading only from the cache and using stale data even when the app is
online and the data served under the URL has changed.

The proposed solution is to fall back to force-cache only on encountering
the first error.