Fix Ref<T> type (#1060)
The previous typing was intended to be:
```typescript
 (instance: T | null) => void |
 RefObject<T> |
 null;
```

But because of a lack of parenthesis, the actual type was:
```typescript
 (instance: T | null) => (void | RefObject<T> | null);
```
which breaks the type if you're using RefObject. Since I'd be updating the type anyway, I just copied the existing type from @react/types which also includes a hack to enable bivariance on callback refs.